Fix bug where non-PASV transfers failed.
On Linux (and possibly other systems), an IPv6 socket can accept a
connection from what's really an IPv4 address. ftpd assumed that
because the incoming connection was on an IPv6 socket that the
data connection would also use IPv6, but the client sends an IPv4
address. The connect() call fails (with errno == 0!) because of
the IPv4/IPv6 mismatch.

This would probably not be a problem if connect() would accept a
struct sockaddr_in* for an IPv6 socket.