Try looking up SIDs as a user first.
This is better than the previous approach of looking up as SID as a group
because user objects would have a gid attribute. Looking up a group SID as 
a user causes an early failure (no uid attribute). This is to cater
for VAS implementations that do not check the objectClass attribute.


git-svn-id: svn+ssh://slccvsl02.prod.quest.corp/opt/svn/rc/vasidmap/trunk@72 d2f197e3-5317-0410-bdde-f57cf7c060d4