Check in 1.0.11.
Changes in ChangeLog. 

This version is tring to add s390x support, but it is not working. Something is different in the s390x version of the same DB2 UDB version. It thinks my code is invalid ( error 2 on SQL1366 I think, plugin type invalid ). 

TODO:
Add tests for each provided section ( Server, Client, and groups ).

Seperate the tests into the above as well. 

Figure out and finish s390x OS.